+++ agent/lib/include/agentxx/agent/agent_io_transport.h
@@ -65,8 +65,8 @@ struct WireTurnResult {
     bool        hasError = false;
     std::string errorMessage;
     bool        interrupted = false;
-    int32_t     startTimeMs = 0; // 轮次开始时间戳 (毫秒)
-    int32_t     durationMs  = 0; // 运行时长 (毫秒)
+    int64_t     startTimeMs = 0; // 轮次开始时间戳 (毫秒)
+    int64_t     durationMs  = 0; // 运行时长 (毫秒)
 };
 
 struct WireContextStats {
